Train Ticket Finder
Sometimes it can be hard to find an empty seat on high-speed trains. The program locks a seat from the selected train autonomously. If there is not any available seat, this program refreshes the ticket page every 20 seconds and
searches for an empty seat. When it finds an available seat, it locks the seat by selecting it.
Python selenium.